fixed numeric truncation error

+++ src/include/ndpi_api.h
@@ -1698,7 +1698,7 @@ extern "C" {
   void ndpi_init_data_analysis(struct ndpi_analyze_struct *s, u_int16_t _max_series_len);
   void ndpi_free_data_analysis(struct ndpi_analyze_struct *d, u_int8_t free_pointer);
   void ndpi_reset_data_analysis(struct ndpi_analyze_struct *d);
-  void ndpi_data_add_value(struct ndpi_analyze_struct *s, const u_int32_t value);
+  void ndpi_data_add_value(struct ndpi_analyze_struct *s, const u_int64_t value);
 
   /* Sliding-window only */
   float ndpi_data_window_average(struct ndpi_analyze_struct *s);
